==> bioNLP 2/1_bioNLP_Data.ipynb <==
!sudo apt install tesseract-ocr
!sudo apt-get install poppler-utils 
!pip install pytesseract pdf2image zero-shot-re neo4j

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

cd /content/drive/MyDrive/

# !git clone 'https://github.com/sudha-vijayakumar/BERN2_TigerGraph_BioMedical_KnowledgeGraph.git'

cd /content/drive/MyDrive/BERN2_TigerGraph_BioMedical_KnowledgeGraph/bioNLP 2/

## Parse pdf for sentences

import requests
import pdf2image
import pytesseract

pdf = '2111.09414.pdf'
doc = pdf2image.convert_from_path(pdf)


article = []
for page_number, page_data in enumerate(doc):
    txt = pytesseract.image_to_string(page_data).encode("utf-8")
    if page_number < 6:
      article.append(txt.decode("utf-8"))
article_txt = " ".join(article)


article_txt

## Text preprocessing

import nltk
nltk.download('punkt')

def clean_text(text):
  """Remove section titles and figure descriptions from text"""
  clean = "\n".join([row for row in text.split("\n") if (len(row.split(" "))) > 3 and not (row.startswith("(a)"))
                    and not row.startswith("Figure")])
  return clean

text = article_txt.split("Abstract")[1]
ctext = clean_text(text)
sentences = nltk.tokenize.sent_tokenize(ctext)

import hashlib
import requests

#http://bern2.korea.ac.kr/documentation
def query_raw(text, url="http://bern2.korea.ac.kr/plain"):
    return requests.post(url, json={'text': text}).json()


entity_list = []
for s in sentences[:-1]:
  try:
    entity_list.append(query_raw(s))
  except:
    print('invalid sentence')
    
parsed_entities = []
for entities in entity_list:
  e = []
  if not entities.get('annotations'):
    parsed_entities.append({'text':entities['text'], 'text_sha256': hashlib.sha256(entities['text'].encode('utf-8')).hexdigest()})
    continue
  for entity in entities['annotations']:
    other_ids = [id for id in entity['id'] if not id.startswith("BERN")]
    entity_type = entity['obj']
    entity_name = entities['text'][entity['span']['begin']:entity['span']['end']]
    try:
      entity_id = [id for id in entity['id'] if id.startswith("BERN")][0]
    except IndexError:
      entity_id = entity_name
    e.append({'entity_id': entity_id, 'other_ids': other_ids, 'entity_type': entity_type, 'entity': entity_name})
  parsed_entities.append({'entities':e, 'text':entities['text'], 'text_sha256': hashlib.sha256(entities['text'].encode('utf-8')).hexdigest()})

### Parse entities

import json
with open('article.txt', 'w') as f:
    f.write(article_txt)


json_string = json.dumps(parsed_entities)
with open('parsed_entities.json', 'w') as outfile:
    outfile.write(json_string)

### Parse Relationships

from transformers import AutoTokenizer
from zero_shot_re import RelTaggerModel, RelationExtractor

model = RelTaggerModel.from_pretrained("fractalego/fewrel-zero-shot")
tokenizer = AutoTokenizer.from_pretrained("bert-large-uncased-whole-word-masking-finetuned-squad")
relations = ['associated', 'interacts']
extractor = RelationExtractor(model, tokenizer, relations)

import itertools

candidates = [s for s in parsed_entities if (s.get('entities')) and (len(s['entities']) > 1)]
predicted_rels = []
for c in candidates:
  combinations = itertools.combinations([{'name':x['entity'], 'id':x['entity_id']} for x in c['entities']], 2)
  for combination in list(combinations):
    try:
      ranked_rels = extractor.rank(text=c['text'].replace(",", " "), head=combination[0]['name'], tail=combination[1]['name'])
      
      if ranked_rels[0][1] > 0.85:
        predicted_rels.append({'head': combination[0]['id'], 'tail': combination[1]['id'], 'type':ranked_rels[0][0], 'source': c['text_sha256']})
    except:
      pass


json_string = json.dumps(predicted_rels)
with open('predicted_rels.json', 'w') as outfile:
    outfile.write(json_string)
