==> bioNLP 2/2_bioNLP_GetNodesRel.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/BERN2_TigerGraph_BioMedical_KnowledgeGraph/bioNLP 2'

### Review Data

with open('article.txt') as f:
    article_txt = f.read()

article_txt

import json
with open('parsed_entities.json') as f:
    parsed_entities = f.read()
    
parsed_entities = json.loads(parsed_entities)
parsed_entities

author = article_txt.split("\n")[0]
title = " ".join(article_txt.split("\n")[2:4])


print(author)
print(title)

### Create Nodes, Relationships (*.csv) => TigerGraph

import pandas as pd

auth_name = []
auth_id = []

auth_id.append('1')
auth_name.append(author)
data={'id':auth_id,'name':auth_name}
print(data)
df = pd.DataFrame(data)
df.to_csv('author.csv')

article_txt = []
article_id = []

article_id.append('1')
article_txt.append(title)
data={'id':article_id,'text':article_txt}
df = pd.DataFrame(data)
df.to_csv('article.csv')

#auth_id,#article_id
data={'sid':auth_id,'did':article_id}
df = pd.DataFrame(data)
df.to_csv('authWROTE.csv')

import csv
article_id=[]
def create_list_from_json(jsonfile):

    with open(jsonfile) as f:
        data = f.read()
    
    parsed_entities = json.loads(data)
        
    sentences=[]
    sentence_id=[]
    
    entity_id=[]
    entity_name=[]
    other_ids=[]
    entity_type=[]
    
    for s in parsed_entities:

        #Sentence: id,text
        sentences.append(s['text'].replace(',',' '))
        sentence_id.append(s['text_sha256'])
        article_id.append('1')
                
        #Entity: id,name,other_ids,type
        try:
            entities =s['entities']
            for entity in entities:
                entity_id.append(entity['entity_id'])
                entity_name.append(entity['entity_id'])
                other_ids.append(entity['other_ids'])
                entity_type.append(entity['entity_type'])
        except:
            pass

    
    data={'id':sentence_id,'text':sentences}
    df = pd.DataFrame(data)
    df.to_csv('sentences.csv')
    
    data={'sid':article_id,'did':sentence_id}
    df = pd.DataFrame(data)
    df.to_csv('article_sentences.csv')

    data={'id':entity_id,'name':entity_id,'other_ids':other_ids,'type':entity_type}
    df = pd.DataFrame(data)
    df.to_csv('entity.csv')


create_list_from_json('parsed_entities.json')  # create the row to be added to csv for each file (json-file)

with open('predicted_rels.json') as f:
    predicted_rels = f.read()
    
predicted_rels = json.loads(predicted_rels)
print(predicted_rels)

rel_df = pd.DataFrame(predicted_rels)
rel_df.to_csv('relationship.csv')

import pandas as pd
types=rel_df['type'].unique()
id_=[]
for i,n in enumerate(types):
  id_.append(str(i))


data = {'id':id_,'type':types}
df = pd.DataFrame(data)
df.to_csv('rel_type.csv')

df
