==> bioNLP 2/3_bioNLP2graph.ipynb <==
!pip3 install pyTigerGraph

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/BERN2_TigerGraph_BioMedical_KnowledgeGraph/bioNLP 2'

import pyTigerGraph as tg
import json
import pandas as pd

hostName = "https://mvie.i.tgcloud.io"
userName = "tigergraph"
password = "tigergraph"

conn = tg.TigerGraphConnection(host=hostName, username=userName, password=password)

print("Connected")

### Clear the Whole Graph
DANGER ZONE

# conn.gsql('''
# USE GLOBAL
# DROP ALL
# ''')

### Create Schema

conn.gsql('''
CREATE VERTEX Sentence(PRIMARY_ID id STRING, text STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE VERTEX Author(PRIMARY_ID id STRING, name STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE VERTEX Entity(PRIMARY_ID id STRING, name STRING, other_ids STRING, type STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE VERTEX Research_Paper(PRIMARY_ID id STRING, title STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE VERTEX Type(PRIMARY_ID type STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE DIRECTED EDGE relation(FROM Entity, TO Type) WITH REVERSE_EDGE="reverse_relation"
CREATE DIRECTED EDGE has_entity(FROM Type, TO Entity) WITH REVERSE_EDGE="reverse_has_entity"
CREATE DIRECTED EDGE has_bio_entity(FROM Sentence, TO Entity) WITH REVERSE_EDGE="reverse_has_bio_entity"
CREATE DIRECTED EDGE contains(FROM Research_Paper, TO Sentence) WITH REVERSE_EDGE="reverse_contains"
CREATE DIRECTED EDGE authored(FROM Author, TO Research_Paper) WITH REVERSE_EDGE="reverse_authored"


''')

### Add Index

conn.gsql('''
CREATE GLOBAL SCHEMA_CHANGE JOB global_schema_change_303694980 {
      ALTER VERTEX Sentence ADD INDEX text_q0sexm9h0od ON (text);
      ALTER VERTEX Author ADD INDEX name_b9v9xwpkoxe ON (name);
      ALTER VERTEX Entity ADD INDEX name_mbd9zqa5dgf ON (name);
      ALTER VERTEX Article ADD INDEX title_4x42wkkn972 ON (title);
      ALTER VERTEX Relation ADD INDEX type_q0v48u0xfqm ON (type);
    }

CREATE GLOBAL SCHEMA_CHANGE JOB global_schema_change {
      ALTER VERTEX Sentence ADD INDEX text_q0sexm9h0od ON (text);
      ALTER VERTEX Author ADD INDEX name_b9v9xwpkoxe ON (name);
      ALTER VERTEX Entity ADD INDEX name_mbd9zqa5dgf ON (name);
      ALTER VERTEX Article ADD INDEX title_4x42wkkn972 ON (title);
      ALTER VERTEX Relation ADD INDEX type_q0v48u0xfqm ON (type);
    }
''')

### Create Graph, Load JOBS

conn.gsql('''
      

CREATE GRAPH bio(Sentence, Author, Entity, Research_Paper, Type, relation, has_entity, has_bio_entity, contains, authored)
set exit_on_error = "false"
CREATE LOADING JOB load_job_article_csv_1647329010113 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX Research_Paper VALUES($1, $2) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE LOADING JOB load_job_article_sentences_csv_1647329019446 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO EDGE contains VALUES($1, $2) USING SEPARATOR=",", HEADER="true", EOL="\n";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE LOADING JOB load_job_authWROTE_csv_1647329024269 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO EDGE authored VALUES($1, $2) USING SEPARATOR=",", HEADER="true", EOL="\n";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE LOADING JOB load_job_author_csv_1647329032144 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX Author VALUES($1, $2) USING SEPARATOR=",", HEADER="true", EOL="\n";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE LOADING JOB load_job_entity_csv_1647329040312 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX Entity VALUES($1, $2, $3, $4) USING SEPARATOR=",", HEADER="true", EOL="\n";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE LOADING JOB load_job_rel_type_csv_1647329053478 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX Type VALUES($2) USING SEPARATOR=",", HEADER="true", EOL="\n";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE LOADING JOB load_job_relationship_csv_1647329061926 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO EDGE has_bio_entity VALUES($4, $1) USING SEPARATOR=",", HEADER="true", EOL="\n";
      LOAD MyDataSource TO EDGE has_bio_entity VALUES($4, $2) USING SEPARATOR=",", HEADER="true", EOL="\n";
      LOAD MyDataSource TO EDGE has_entity VALUES($3, $1) USING SEPARATOR=",", HEADER="true", EOL="\n";
      LOAD MyDataSource TO EDGE relation VALUES($2, $3) USING SEPARATOR=",", HEADER="true", EOL="\n";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE LOADING JOB load_job_sentences_csv_1647329071260 FOR GRAPH bio {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX Sentence VALUES($1, $2) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
    }

set exit_on_error = "true"
set exit_on_error = "false"
CREATE QUERY get_disease_relation(/* Parameters here */) FOR GRAPH bio { 
  /* Write query logic here */ 
  ListAccum<EDGE> @relation;

  start = {Entity.*};

  entity_typ = SELECT s FROM start:s -(relation:r)- :t
               ACCUM s.@relation+=r;

  PRINT entity_typ;
}
set exit_on_error = "true"

''')

## Fetch Access Token

graphName = "bio"
conn.graphname = graphName
secret = conn.createSecret()
token = conn.getToken(secret, setToken=True)

### Peek into Data to load

df = pd.read_csv('article.csv')
print(df.head(0))
print('######')
df = pd.read_csv('author.csv')
print(df.head(0))
print('######')
df = pd.read_csv('entity.csv')
print(df.head(0))
print('######')
df = pd.read_csv('sentences.csv')
print(df.head(0))
print('######')
df = pd.read_csv('relationship.csv')
print(df.head(0))
print('######')
df = pd.read_csv('authWROTE.csv')
print(df.head(0))
print('######')
df = pd.read_csv('rel_type.csv')
print(df.head(0))
print('######')
df = pd.read_csv('article_sentences.csv')
print(df.head(5))
print('######')

### Load data

article_file = "article.csv"
res = conn.uploadFile(article_file, fileTag='MyDataSource', jobName='load_job_article_csv_1647329010113')
print(json.dumps(res, indent=2))

author_file = "author.csv"
res = conn.uploadFile(author_file, fileTag='MyDataSource', jobName='load_job_author_csv_1647329032144')
print(json.dumps(res, indent=2))

entity_file = "entity.csv"
res = conn.uploadFile(entity_file, fileTag='MyDataSource', jobName='load_job_entity_csv_1647329040312')
print(json.dumps(res, indent=2))

sentences_file = "sentences.csv"
res = conn.uploadFile(sentences_file, fileTag='MyDataSource', jobName='load_job_sentences_csv_1647329071260')
print(json.dumps(res, indent=2))

rel_file = "relationship.csv"
res = conn.uploadFile(rel_file, fileTag='MyDataSource', jobName='load_job_relationship_csv_1647329061926')
print(json.dumps(res, indent=2))

aw_file = "authWROTE.csv"
res = conn.uploadFile(aw_file, fileTag='MyDataSource', jobName='load_job_authWROTE_csv_1647329024269')
print(json.dumps(res, indent=2))

df = pd.read_csv("article_sentences.csv")
df

as_file = "article_sentences.csv"
res = conn.uploadFile(as_file, fileTag='MyDataSource', jobName='load_job_article_sentences_csv_1647329019446')
print(json.dumps(res, indent=2))

rel_file = "rel_type.csv"
res = conn.uploadFile(aw_file, fileTag='MyDataSource', jobName='load_job_rel_type_csv_1647329053478')
print(json.dumps(res, indent=2))
